+++ http/src/main/scala/com/twitter/finatra/modules/package.scala
@@ -1,9 +1,0 @@
-package com.twitter.finatra
-
-package object modules {
-  @deprecated("Use com.twitter.finatra.http.modules.AccessLogModule", "")
-  val AccessLogModule = com.twitter.finatra.http.modules.AccessLogModule
-
-  @deprecated("Use com.twitter.finatra.http.modules.MustacheModule", "")
-  val MustacheModule = com.twitter.finatra.http.modules.MustacheModule
-}